src/requirements.txt
dukpy==0.5.0
skia-python==138.0
pysdl2==0.9.17
pysdl2-dll==2.32.0
PyOpenGL==3.1.10